pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sre-observability"
version = "0.1.0"
description = "Unified Prometheus & OpenTelemetry observability for internal Python services"
requires-python = ">=3.9"
dependencies = [
    "prometheus-client>=0.19",
    "psutil>=5.9",
    "opentelemetry-api>=1.22",
    "opentelemetry-sdk>=1.22",
    "opentelemetry-exporter-otlp-proto-grpc>=1.22",
    "opentelemetry-exporter-prometheus>=0.43b0",
    "opentelemetry-instrumentation-logging>=0.43b0",
]

[project.optional-dependencies]
fastapi = [
    "fastapi",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
]
flask = [
    "flask",
    "opentelemetry-instrumentation-flask>=0.43b0",
]
requests = [
    "opentelemetry-instrumentation-requests>=0.43b0",
]
all = [
    "sre-observability[fastapi,flask,requests]",
]

[tool.setuptools.packages.find]
where = ["src"]
